+++ backend/controllers/CreditosController.php
@@ -38,6 +38,7 @@ class CreditosController extends Controller
      */
     public function actionIndex()
     {
+        if(!Yii::$app->user->isGuest && Yii::$app->user->identity->role_id >= 30){
         $searchModel = new CreditosSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -45,6 +46,10 @@ class CreditosController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else{
+            return $this->redirect(['site/permiso']);
+            //return $this->render('site/nopermitido');
+       }
     }
 
     /**
@@ -54,9 +59,14 @@ class CreditosController extends Controller
      */
     public function actionView($id)
     {
+         if(!Yii::$app->user->isGuest && Yii::$app->user->identity->role_id >= 30){
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        }else{
+            return $this->redirect(['site/permiso']);
+            //return $this->render('site/nopermitido');
+       }
     }
 
     /**
@@ -66,6 +76,7 @@ class CreditosController extends Controller
      */
     public function actionCreate()
     {
+         if(!Yii::$app->user->isGuest && Yii::$app->user->identity->role_id >= 30){
         $imagen="";
         $model= new Creditos;
         if ($model->load(Yii::$app->request->post())) {
@@ -87,6 +98,11 @@ class CreditosController extends Controller
                 'model' => $model,
             ]);
         }
+
+        }else{
+            return $this->redirect(['site/permiso']);
+            //return $this->render('site/nopermitido');
+       }
     }
 
     /**
@@ -97,6 +113,8 @@ class CreditosController extends Controller
      */
     public function actionUpdate($id)
     {
+
+        if(!Yii::$app->user->isGuest && Yii::$app->user->identity->role_id >= 30){
          $model = $this->findModel($id);
         
         if ($model->load(Yii::$app->request->post()) ) {
@@ -124,6 +142,11 @@ class CreditosController extends Controller
                 'model' => $model,
             ]);
         }
+
+        }else{
+            return $this->redirect(['site/permiso']);
+            //return $this->render('site/nopermitido');
+       }
     }
 
     /**
@@ -134,9 +157,14 @@ class CreditosController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->isGuest && Yii::$app->user->identity->role_id >= 30){
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        }else{
+            return $this->redirect(['site/permiso']);
+            //return $this->render('site/nopermitido');
+       }
     }
 
     /**
